openssl: make it easier to build without the `*-pdb` package
In Git for Windows, we generate and bundle the `.pdb` files of OpenSSL.

However, this requires Visual Studio to be installed (some private
`.dll` files from that software are used to generate those `.pdb` files,
after all). So for contributors who do not have Visual Studio installed,
the `mingw-w64-openssl` package simply does not build.

A simple workaround would be to remove the `*-pdb` package from the
package list in `PKGBUILD`. With this patch, this works.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>