ca-certificates: sync with MSYS2
There is a very broken mingw-w64-x86_64-ca-certificates 20180409-4 that
does not even allow fetching from GitHub (!).

To fix this, the package was rebuilt and installed into the repositories
at https://github.com/git-for-windows/git-sdk-64 (and -32).

Should this become a more permanent situation, we might want to use the
post-install script of `git-extra` to download the `ca-bundle.crt` file
from those repositories if necessary.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>