git: avoid problems with the sha1collisiondetection submodule
It is unfortunately a known problem that `git stash` cannot handle
submodules well (see t/t3906-stash-submodule.sh).

Which means that the code to stash the "changes" when trying to check
out the source code with LF-only line endings does not work well,
failing with an error message indicating that one cannot stage an empty
directory.

Let's ignore the bug in `git stash` for now, and replace the call by
`git reset --hard`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>