git: handle return value 0 of GetFinalPathNameByHandleW()
When `GetFinalPathNameByHandleW()` returns 0, it means that it failed.
In this case, we want fall back to using the originally-provided
executable path, even if it might contain unresolved symbolic links.

However, we did not fall back to using this path when that function call
failed, except when it failed in a way that indicates that the resulting
path would be too long.

The symptom of this bug is a rather obscure:

	Invalid executable path:

This commit addresses that, fixing
https://github.com/git-for-windows/git/issues/2493

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>