curl: prepare for v7.69.0 and add schannel revoke-best-effort
With this change, we rebase the two patches we have on top of cURL to
the newest cURL version, and we also add the (still tentative) support
for the "best effort" strategy for dealing with revoke problems in the
Secure Channel backend. This "best effort" strategy is currently under
review at https://github.com/curl/curl/pull/4981.

We will use this feature in Git for Windows, to turn on "best-effort" by
default (using `CURLSSLOPT_REVOKE_BEST_EFFORT` as a tell-tale whether
cURL supports it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>