git: move the mingit-busybox-test-artifacts Makefile rule
The `install-mingit-test-artifacts` target was introduced solely for the
benefit of `mingw-w64-git`, and it lives in Git for Windows' `Makefile`
so far.

This is not desirable, though, as it makes it impossible to use the very
same build procedure to build packages for testing from, say, upstream
Git's `pu` branch.

Let's move the target into the `mingw-w64-git.mak` file instead, so that
we can target more Git revisions, especially from upstream Git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>