git: install /bin/{git,bash,sh}
For backward compatibility, the Git for Windows installer provides a
/bin/ directory with `git.exe`, `bash.exe` and `sh.exe` in it. These are
mere redirectors (like `/cmd/git.exe`) for the benefit of letting users
run these executables from native Win32 programs without having to worry
about setting the correct environment variables to the correct values.

Let's face it: we will never be able to get rid of those files in the
`/bin/` directory. Yes, they are only there for compatibility. Yes,
MSYS2 will hide them by overlaying `/bin/` with the contents of
`/usr/bin/`. And yes, they do come in useful when running things e.g.
via PowerShell.

So let's not wait for the Git for Windows installer to put them into
place, but let's let the mingw-w64-git package already do so.

Since `/bin/` is overlayed, we have to play a trick here: we use the
absolute Windows path so that we still can use `cp`. We also have to do
that in the post-install function, and a consequence is that these files
are not marked as "owned by mingw-w64-git".

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>