git: include the edit-git-bash helper
This helper allows to configure what executable to call, with what
options, and it was included only in Git for Windows' installer, but
the mingw-w64-git package is a much more logical place.

Please note that this is copy/edited from
https://github.com/git-for-windows/build-extra/blob/e97e777dbeea/edit-git-bash.c

The only edit: The `#ifdef STANDALONE_EXE` guards were removed, in favor
of _only_ building an .`exe` (and no longer a `.dll` as we did
originally, to call the function directly from within the
InnoSetup-generated installer).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>