git: introduce experimental support for /etc/git-bash.config
Currently, this support consists on parsing that file for lines of the
form `MSYS=...`. If found, the environment variable `MSYS` will be set
accordingly before launching `git-bash.exe` (or `git-cmd.exe` or
`git-gui.exe`). If the environment variable `MSYS` is already set, the
value from the file will be prepended (so that the user-set environment
variable settings still take precedence, i.e. if the file says
`MSYS=disable_pcon` and the environment variable says `nodisable_pcon`,
the latter will "win").

This feature is experimental, for the time being.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>