git: do not let `edit-git-bash.exe` depend on `libssp-0.dll`
With this change, we closely imitate the way `git-bash.exe` is built:
with minimal dependencies. That is needed so that `edit-git-bash.exe`
can be called from within Git for Windows' installer via its absolute
path (and without adjusting `PATH` so that it includes Git for Windows'
`/mingw64/bin`).

This fixes https://github.com/git-for-windows/git/issues/2631

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>