curl: mention that brotli is a dependency
The binary does already depend on libbrotli, but the Pacman package's
metadata did not spell it out. Therefore, `pacman -S
mingw-w64-x86_64-curl` would not ensure that the dependency is
installed, and Git for Windows' scripts would not include it in the
installer, Portable Git nor MinGit.

This fixes https://github.com/git-for-windows/git/issues/2737.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>