git: add ARM64 runtime detection
In order to support native Git for Windows ARM64 binaries,
this commit introduces ARM64 runtime detection. Git-wrapper itself will
remain an i686 or x86_64 executable for now, but since Windows on ARM
has built-in i686 emulation, it can run the file.

If an ARM64 system is detected, we set MSYSTEM=ARM64. Next to that,
because many binaries like curl aren't available natively for ARM64 yet,
we add /mingw32/bin to the path, so that we'll fallback to i686
executables in that case.

Signed-off-by: Tommy Vercetti <7903172+tommyvct@users.noreply.github.com>
Signed-off-by: Dennis Ameling <dennis@dennisameling.com>