git: ignore HOMEDRIVE/HOMEPATH default for non-shell logons
Internally, Git expects the environment variable `HOME` to be set, and
to point to the current user's home directory.

This environment variable is not set by default on Windows, and
therefore Git tries its best to construct one if it finds `HOME` unset.

There are actually two different approaches Git tries: first, it looks
at `HOMEDRIVE`/`HOMEPATH` because this is widely used in corporate
environments with roaming profiles, and a user generally wants their
global Git settings to be in a roaming profile.

Only when `HOMEDRIVE`/`HOMEPATH` is either unset or does not point to a
directory, Git will fall back to using `USERPROFILE` instead.

However, starting with Windows Vista, for secondary logons (think:
`runas.exe /user:<user>` and services, the environment variables
`HOMEDRIVE`/`HOMEPATH` point to Windows' system directory (usually
`C:\Windows\system32`).

That is undesirable, and that location is usually write-protected
anyway.

Let's detect that situation and ignore `HOMEDRIVE`/`HOMEPATH` in that
instance, forcing the fall-back to `USERPROFILE`.

This addresses the second half of
https://github.com/git-for-windows/git/issues/2709

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>