git: teach the wrapper to stand in for `scalar.exe`
We already include `scalar.exe` in the mingw-w64-git package (only
if `contrib/scalar/` exists in the Git checkout, of course).

However, the standard way to install Git for Windows is to leave its
`bin/` directory out of the `PATH`, instead adding `cmd/`. That
directory contains only a few executables, among them `git.exe`,
`git-gui.exe` and `gitk.exe` (which are copies of the Git wrapper which
does nothing else than modify the environment a bit and then hand off to
the _real_ executable in `/mingw64/bin/`)).

Therefore, we need to provide the same functionality for `scalar.exe` in
the `<Git>/cmd/` directory, right next to those other copies of the Git
wrapper.

This change makes it possible for the Git wrapper to act as
redirector for `scalar.exe`, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>