git: if the source code for `scalar.exe` was found, include it
The Scalar project at https://github.com/microsoft/scalar set out to
apply the lessons from VFS for Git to allow scaling Git to large
monorepos. The difference is that Scalar wants to make Git _aware_ of
the sparseness, which allows it to drop the requirement for a virtual
file system driver.

Most of the features developed in Scalar have been ported to C and been
included in Git proper:

- partial clones
- sparse checkouts
- cone mode of sparse checkouts
- commit graphs
- multi-pack index
- scheduled maintenance

The remainder seems to be split into two categories:

- opinionated defaults
- file system monitor
- support for Azure Repos (which does not support partial clones)

Naturally, we want to port that remaining functionality to C, too, and
microsoft/git already has the beginnings of `contrib/scalar/`.

With this commit, we include `scalar.exe` if `contrib/scalar/` is
present in the source code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>