Revert "Merge pull request #48 from dscho/scalar-in-gits-contrib"
After reflecting on the design of `contrib/scalar/`, I've revamped it
(twice, actually). And I settled on a better solution (both in terms of
convenience as well as in terms of intrusiveness): now all one has to do
to include Scalar is to set `INCLUDE_SCALAR=YesPlease`. That flag is
handled in Git's `Makefile`, and (almost) no change is required on
`mingw-w64-git`'s side.

The only thing we _will_ need is to adjust the Git wrapper to stand in
for `scalar.exe`, and that will be done in the next commit.

This reverts commit ef41cf9e0 (Merge pull request #48 from
dscho/scalar-in-gits-contrib, 2021-04-29), i.e. reversing the patches
merged in https://github.com/git-for-windows/MINGW-packages/pull/48.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>