git: let the Git wrapper stand in for git-remote-http
The `git-remote-http.exe` and `git-remote-https.exe` files are hardlinks
of one another.

In scenarios where hardlinks are not supported (as is the case in `.zip`
files, e.g. in MinGit) this means: a copy of a 2MB file.

Let's teach the Git wrapper to serve as a redirector in place of
`git-remote-http.exe`, to save those two megabytes.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>