git: teach the Git wrapper to redirect `sh.exe` to BusyBox' ash
This is another building block required to make the BusyBox-flavor of
MinGit fully functional: OpenSSH insists on having a shell to execute
any configured `ProxyCommand`, and we do have one: BusyBox' ash. But
OpenSSH insists on having it as an executable to which `SHELL` points,
without having to specify a specific first argument.

So let's teach the Git wrapper the trick to hand off to BusyBox if it is
installed into `/mingw64/bin/sh.exe` (or `ash.exe`, or into `/bin` or
into `/mingw64/libexec/busybox/`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>