Add a GitHub workflow to build and deploy a single MINGW package
This is the equivalent to
https://github.com/git-for-windows/MSYS2-packages/blob/main/.github/workflows/build-and-deploy.yml
which Git for Windows had to introduce in order to provide backports for
i686 variants of the MSYS packages (MSYS2 no longer maintains those).

This was developed in case we needed to build and deploy a MINGW package
manually to fix a bug in mingw-w64-gettext that prevented Git for
Windows v2.35.0-rc0 from being released.

In the meantime, MSYS2 accepted the PR and we do not have to build this
ourselves.

But it will provide a good starting point when we start transitioning
Git for Windows' packaging process away from Azure Pipelines toward
GitHub workflows (and probably also away from doing everything in
`/usr/src/build-extra/please.sh upgrade <package>` toward a more
PR-centric flow).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>