git-wrapper: cache less of is_running_on_arm64()
is_running_on_arm64() currently checks two distinct things:
 - is this process running on ARM64 hardware?
 - is there a clangarm64/bin directory in the specified parent directory?

While the answer to the first question won't change during execution,
the answer to the second question can change depending on function
parameters. We currently cache the result of both in one variable.
Due to this, once we checked in one directory, we're ignoring all
subsequent directories.

This causes /bin/sh.exe and /bin/bash.exe to fail with the message
"Top-level not found" on ARM64.

To fix this, split is_running_on_arm64() into two parts and cache only
the result of the hardware part.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>