openssl: fix 05-test_rand
As of OpenSSL v3.1.4, there is a pair of new test failures in
05-test_rand. The symptom looks like this:

	Engine "ossltest" set.
	../../util/wrap.pl ../../apps/openssl.exe rand -engine ossltest -hex 16 => 0
	not ok 4 - rand with ossltest: Check rand output is as expected

	#   Failed test 'rand with ossltest: Check rand output is as expected'
	#   at ../openssl-3.1.4/test/recipes/05-test_rand.t line 32.
	Engine "dasync" set.
	../../util/wrap.pl ../../apps/openssl.exe rand -engine dasync -hex 16 => 0
	not ok 5 - rand with dasync: Check rand output is of expected length

The two failing tests both spawn an engine, which by virtue of being
MINGW `.dll` files use CR/LF when printing to `stdout`. To accommodate
for that, use the same "better chomp" as elsewhere in OpenSSL's source
code.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>