curl: fix `--write-out '%{http_code}'` with Secure Channel
The `git update-git-for-windows` command is a Unix Shell script that
relies heavily on `curl`; For example, it calls:

  curl --silent --show-error --output /tmp/gfw-httpget-ruMLqjGG.txt \
    --write-out '%{http_code}' \
    https://api.github.com/repos/git-for-windows/git/releases/latest

This worked well until cURL v8.8.0 was integrated. With the Secure
Channel backend, the invocation would fail with this error message:

  curl: (43) A libcurl function was given a bad argument

The same issue had been reported for a different SSL backend:
https://github.com/curl/curl/issues/13845. Conveniently, all we have to
do is to Backport the fix from the PR that addressed that ticket to
address the bug at hand.

This fixes https://github.com/microsoft/git/issues/655.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>