git: force-fully ignore the environment variable `GIT_VERSION`
While it used to be possible to override the version shown in `git
version` via `make GIT_VERSION=...`, the same was not true of the
environment variable `GIT_VERSION`.

However, that changed in Git's 992bc5618f2 (GIT-VERSION-GEN: fix
overriding GIT_VERSION, 2024-12-20).

To force the build process to derive the version shown in the output of
`git version`, let's forcefully un-set the environment variable during
the build.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>