git: drop the -test-artifacts package
It looked like a good idea at the time to ship test artifacts alongside
the main package so as to be able to test MinGit-busybox. In practice,
we never got BusyBox production-ready (and won't, given that there is no
appetite in the BusyBox-w32 project to introduce even basic things like
Continuous Integration). Therefore, building the `-test-artifacts`
package just increases the build time and artifacts size without any
benefit.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>