git: drop the commit hash from pkgver for MSYS2 compatibility
The practice of appending the short commit hash to the package version
was useful for Git for Windows SDK development to track exactly which
commit was used for a given build. However, this is not standard
practice in MSYS2 and seems inappropriate for upstreaming this package
definition.

Remove the `pkgver()` function that dynamically added the commit hash.
The `pkgver` is now derived from the tag variable using shell parameter
expansion: `${tag%.windows.*}.${tag##*.windows.}` converts
`2.51.2.windows.1` to `2.51.2.1`.

While at it, no longer skip verifying the exact commit hash. This
requires us to force `core.autoCRLF` because `makepkg`'s checksum
algorithm is unstable, producing diverging checksums depending on that
config setting.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>