git: split out git-subtree into separate package
The `git subtree` command is not actually a canonical Git command
(carried only in the `contrib/` tree, with no clear path into the core
commands). Git for Windows ships this command to be
backwards-compatible, as it has been distributed with Git for Windows
for a long time already.

Let's split it out into its own package, a first step towards
streamlining the `mingw-w64-git` package so that it contains only the
actively-supported (or at least "official") Git commands.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>