git: split out Git for Windows' extras into a separate package
Git for Windows' `mingw-w64-git` package currently includes not only the
canonical Git build artifacts, but also includes the "Git wrapper" and
other files that are useful in Git for Windows' context, but would be
inappropriate in the canonical MSYS2 version of `mingw-w64-git`.

The current effort to align Git for Windows better with MSYS2, most
notably including the effort to upstream as much of Git for Windows as
possible, is kind of blocked by this. For example, in regular MSYS2, it
should be no problem to install i686, x86_64 and clang-aarch64 versions
side by side, but e.g. `/git-bash.exe` would prevent that from being
possible.

It is easy to solve this conundrum, though, by splitting out those Git
for Windows add-ons into a separate package.

This commit is best viewed with `--color-moved`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>