git: split out Git.pm into its own package
The only remaining Perl-dependent part in the `mingw-w64-git` package is
the Perl module ("Git.pm"). Let's split it out into the new
`mingw-w64-git-perl` package, adding it as a dependency to the
previously-split out packages containing the Git commands that depend on
that Perl module.

Note: The `install-perl-module` rule installs not only `Git.pm` but also
the `Git::SVN` module. However, that module is already owned by
`mingw-w64-git-svn` and that needs to remain so.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>