git: split out the Perforce helper into a separate package
This is the only Python script in Git, so after splitting it into its
own package, `mingw-w64-git` can stop depending on `python`.

Technically, this new package should be a dependency of
`mingw-w64-git-for-windows-addons` so that the `git-p4` script is
included in Git for Windows again. However, while that script had been
included, it was not functional because no Python interpreter is
distributed with Git for Windows. So it's better to omit it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>